+++ src/Hacks/Cosmetic/NoOrbPulse.cpp
@@ -1,0 +1,34 @@
+#include "../../Client/Module.hpp"
+#include <Geode/modify/PlayLayer.hpp>
+
+using namespace geode::prelude;
+
+class NoOrbPulse : public Module
+{
+    public:
+        MODULE_SETUP(NoOrbPulse)
+        {
+            setName("No Orb Pulse");
+            setID("no-orb-pulse");
+            setCategory("Cosmetic");
+            setDescription("Disables orb's from pulsing, Doesn't work on main levels made before 2.2");
+        }
+};
+
+SUBMIT_HACK(NoOrbPulse);
+
+class $modify (PlayLayer)
+{
+    virtual void updateVisibility(float dt)
+    {
+        auto fmod = FMODAudioEngine::get();
+        auto pulse1 = fmod->m_pulse1;
+
+        if (NoOrbPulse::get()->getRealEnabled())
+            fmod->m_pulse1 = 1.0f;
+
+        PlayLayer::updateVisibility(dt);
+
+        fmod->m_pulse1 = pulse1;
+    }
+};