+++ src/Labels/Hooks.hpp
@@ -2,8 +2,16 @@
 
 #include <Geode/Geode.hpp>
 #include <Geode/modify/PlayLayer.hpp>
+#include <Geode/modify/GJBaseGameLayer.hpp>
 
 class $modify (LabelPlayLayer, PlayLayer)
 {
     bool init(GJGameLevel* level, bool useReplay, bool dontCreateObjects);
+
+    void resetLevel();
+};
+
+class $modify (LabelBaseGameLayer, GJBaseGameLayer)
+{
+    void handleButton(bool down, int button, bool isPlayer1);
 };